+++ tests/index.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable sonarjs/no-identical-functions */
 import helpers, { RunResult } from 'yeoman-test';
 import path from 'path';
 import { exec } from 'child_process';
@@ -16,6 +17,45 @@ describe('Generator', () => {
           runResult = result;
           done();
         });
+    });
+    it('Run build', (done) => {
+      exec('yarn build', { cwd: path.resolve(runResult.cwd, projectName) }, (error, stdout) => {
+        expect(error).toBeNull();
+        expect(stdout).not.toBeNull();
+        done();
+      });
+    });
+    it('Run test', (done) => {
+      exec('yarn test', { cwd: path.resolve(runResult.cwd, projectName) }, (error, stdout) => {
+        expect(error).toBeNull();
+        expect(stdout).not.toBeNull();
+        done();
+      });
+    });
+  });
+  describe('Recoil generator', () => {
+    let runResult: RunResult;
+    const projectName = 'recoil-project';
+    beforeAll((done) => {
+      return helpers
+        .run(path.join(__dirname, '../app'))
+        .inDir(path.resolve(__dirname, 'apps/recoil'))
+        .withPrompts({
+          state: 'recoil'
+        })
+        .withArguments([projectName])
+        .withOptions({ 'no-update': true })
+        .then((result) => {
+          runResult = result;
+          done();
+        });
+    });
+    it('Run build', (done) => {
+      exec('yarn build', { cwd: path.resolve(runResult.cwd, projectName) }, (error, stdout) => {
+        expect(error).toBeNull();
+        expect(stdout).not.toBeNull();
+        done();
+      });
     });
     it('Run test', (done) => {
       exec('yarn test', { cwd: path.resolve(runResult.cwd, projectName) }, (error, stdout) => {