+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.4.3](https://github.com/mrmilu/generator-react-mrmilu/compare/v1.4.2...v1.4.3) (2021-03-10)
+
+
+### Bug Fixes
+
+* update recoil sample ([eb7b014](https://github.com/mrmilu/generator-react-mrmilu/commit/eb7b01426e6c7e17bcacc6ed413469710f75fe47))
+
+## [1.4.2](https://github.com/mrmilu/generator-react-mrmilu/compare/v1.4.1...v1.4.2) (2021-03-10)
+
+
+### Bug Fixes
+
+* **json:** update package.json structure ([afb2c5b](https://github.com/mrmilu/generator-react-mrmilu/commit/afb2c5bc91bda23b00d094257e1d3c4d5d3473f4))
+
 ## [1.4.1](https://github.com/mrmilu/generator-react-mrmilu/compare/v1.4.0...v1.4.1) (2021-03-10)